Extracts keyword handlers into KeywordHandlers.cpp
